api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.4
